requirements.txt
Jinja2==3.1.1
MarkupSafe==2.1.1
PyPDF2==1.27.6
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
cpe==1.2.1
idna==3.3==3.3
lxml==4.8.0
nvdlib==0.5.6
python-dotenv==0.20.0
requests==2.27.1
setuptools==57.0.0
soupsieve==2.3.2.post1
urllib3==1.26.9
wheel==0.36.2